ci: streamline build phases to skip already-cached stages

+++ build-toolchain.sh
@@ -115,7 +115,11 @@ fi
 
 if [ "x$skip_native_build" != "xyes" ] ; then
     mkdir -p $BUILDDIR_NATIVE
-    rm -rf $INSTALLDIR_NATIVE && mkdir -p $INSTALLDIR_NATIVE
+    if [ -z "$skip_stages" ]; then
+        rm -rf $INSTALLDIR_NATIVE && mkdir -p $INSTALLDIR_NATIVE
+    else
+        mkdir -p $INSTALLDIR_NATIVE
+    fi
     rm -rf $PACKAGEDIR && mkdir -p $PACKAGEDIR
 fi
